+++ ExpenseTracker/App.swift
@@ -14,13 +14,13 @@ class App {
     
     var name: String
     var photo: UIImage?
-    var rating: Int
+    var rating: Double
     var date: String
     
     
     // MARK: Initialization
     
-    init?(name: String, photo: UIImage?, rating: Int, date: String) {
+    init?(name: String, photo: UIImage?, rating: Double, date: String) {
         
         // Initialization should fail if there is no name or if the rating is negative
         if name.isEmpty || rating < 0 {
